+++ src/core/io.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void load_whole_file(const char* path, char** out_data, u64* out_size);
+void load_whole_file(const char* path, char** out_data, int* out_size);
